5주차 2차 반영
